requirements.txt
telethon==1.24.0
pyrogram==2.1.32
python-dotenv
psutil
opencv-python-headless
devgagantools
aiofiles
# ggnpyro
https://www.dl.dropboxusercontent.com/scl/fi/e0fo6fcjn8kmr5r0x6wvg/myownpyro.zip?rlkey=d1znpwckss4ullz0sg7e1qjjg&st=kmbh7wdv&dl=0
aiohttp
telethon-tgcrypto
# MongoDB related packages (optional)
motor
pymongo
pytz
Pillow
tgcrypto
flask
werkzeug==2.2.2
mutagen
yt-dlp
requests
cryptography
python-decouple
ethon==0.1.5
cryptg
pydub
ffmpeg-python
